Make ValidateRequestAttribute, filter, and middleware consistent, improve tests
