+++ src/Twilio.AspNet.Core.UnitTests/TwilioRequestValidationOptionsTests.cs
@@ -1,23 +1,16 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Net;
 using System.Text.Json;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Abstractions;
-using Microsoft.AspNetCore.Mvc.Filters;
-using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
-using Microsoft.Extensions.Primitives;
 using Xunit;
 
 namespace Twilio.AspNet.Core.UnitTests;
 
-public class ValidateRequestAttributeTests
+public class AddTwilioRequestValidationOptionsTests
 {
     private static readonly TwilioOptions ValidTwilioOptions = new()
     {
@@ -127,7 +120,7 @@ public class ValidateRequestAttributeTests
                 new KeyValuePair<string, string>("Twilio", null),
                 new KeyValuePair<string, string>("Twilio:RequestValidation:AuthToken", null)
             }).Build();
-    
+
         serviceCollection.AddSingleton<IConfiguration>(configuration);
         serviceCollection.AddTwilioRequestValidation();
 
@@ -220,153 +213,5 @@ public class ValidateRequestAttributeTests
         var optionsFromDi = await monitor.WaitForOptionChange();
 
         Assert.Equal(options.AuthToken, optionsFromDi.AuthToken);
-    }
-
-    [Fact]
-    public void AddTwilioRequestValidation_Should_Configure_ValidateRequestAttribute()
-    {
-        var serviceCollection = new ServiceCollection();
-        serviceCollection.AddTwilioRequestValidation((_, options) =>
-        {
-            options.AllowLocal = ValidTwilioOptions.RequestValidation.AllowLocal;
-            options.AuthToken = ValidTwilioOptions.RequestValidation.AuthToken;
-            options.BaseUrlOverride = ValidTwilioOptions.RequestValidation.BaseUrlOverride;
-        });
-
-        var serviceProvider = serviceCollection.BuildServiceProvider();
-
-        var attributeFactory = new ValidateRequestAttribute();
-        var attribute =
-            (ValidateRequestAttribute.InternalValidateRequestAttribute)attributeFactory
-                .CreateInstance(serviceProvider);
-
-        Assert.Equal(ValidTwilioOptions.RequestValidation.AllowLocal, attribute.AllowLocal);
-        Assert.Equal(ValidTwilioOptions.RequestValidation.AuthToken, attribute.AuthToken);
-        Assert.Equal(ValidTwilioOptions.RequestValidation.BaseUrlOverride, attribute.BaseUrlOverride);
-    }
-
-    [Fact]
-    public void Creating_ValidateRequestAttribute_Without_AddRequestValidation_Should_Throw()
-    {
-        var serviceCollection = new ServiceCollection();
-        var serviceProvider = serviceCollection.BuildServiceProvider();
-
-        var attributeFactory = new ValidateRequestAttribute();
-        var exception = Assert.Throws<Exception>(() =>
-            (ValidateRequestAttribute.InternalValidateRequestAttribute)attributeFactory
-                .CreateInstance(serviceProvider));
-        Assert.Equal("RequestValidationOptions is not configured.", exception.Message);
-    }
-    
-    [Fact]
-    public async Task ValidateRequestAttribute_Should_Use_Reloaded_Configuration()
-    {
-        const string optionsFile = "ValidateRequestAttributeAutoReload.json";
-        if (File.Exists(optionsFile)) File.Delete(optionsFile);
-        var jsonText = JsonSerializer.Serialize(new { Twilio = ValidTwilioOptions });
-        await File.WriteAllTextAsync(optionsFile, jsonText);
-
-        var serviceCollection = new ServiceCollection();
-        var configuration = new ConfigurationBuilder()
-            .AddJsonFile(optionsFile, optional: false, reloadOnChange: true)
-            .Build();
-
-        serviceCollection.AddSingleton<IConfiguration>(configuration);
-        serviceCollection.AddTwilioRequestValidation();
-
-        var serviceProvider = serviceCollection.BuildServiceProvider();
-        
-        var attributeFactory = new ValidateRequestAttribute();
-        var attribute = (ValidateRequestAttribute.InternalValidateRequestAttribute)attributeFactory
-            .CreateInstance(serviceProvider);
-        
-        Assert.Equal(ValidTwilioOptions.RequestValidation.AuthToken, attribute.AuthToken);
-        Assert.Equal(ValidTwilioOptions.RequestValidation.BaseUrlOverride, attribute.BaseUrlOverride);
-        Assert.Equal(ValidTwilioOptions.RequestValidation.AllowLocal, attribute.AllowLocal);
-
-        
-        TwilioOptions updatedOptions = new()
-        {
-            RequestValidation = new TwilioRequestValidationOptions
-            {
-                AuthToken = "My Twilio:RequestValidation:Updated Auth Token",
-                AllowLocal = true,
-                BaseUrlOverride = "Different URL"
-            }
-        };
-
-        jsonText = JsonSerializer.Serialize(new { Twilio = updatedOptions });
-        await File.WriteAllTextAsync(optionsFile, jsonText);
-        
-        // wait for the option change to be detected
-        var monitor = serviceProvider.GetRequiredService<IOptionsMonitor<TwilioRequestValidationOptions>>();
-        await monitor.WaitForOptionChange();
-
-        // IOptionsSnapshot is calculated per scope
-        using var scope = serviceProvider.CreateScope();
-        attribute = (ValidateRequestAttribute.InternalValidateRequestAttribute)attributeFactory
-            .CreateInstance(scope.ServiceProvider);
-        
-        Assert.Equal(updatedOptions.RequestValidation.AuthToken, attribute.AuthToken);
-        Assert.Equal(updatedOptions.RequestValidation.BaseUrlOverride, attribute.BaseUrlOverride);
-        Assert.Equal(updatedOptions.RequestValidation.AllowLocal,attribute.AllowLocal);
-    }
-
-    [Fact]
-    public void ValidateRequestAttribute_Validates_Request_Successfully()
-    {
-        var attribute = new ValidateRequestAttribute.InternalValidateRequestAttribute(
-            authToken: ContextMocks.fakeAuthToken,
-            null,
-            false
-        );
-
-        var form = new FormCollection(new Dictionary<string, StringValues>
-        {
-            { "key1", "value1" },
-            { "key2", "value2" }
-        });
-        var fakeContext = new ContextMocks(true, form).HttpContext.Object;
-
-        var actionExecutingContext = new ActionExecutingContext(
-            new ActionContext(fakeContext, new RouteData(), new ActionDescriptor()),
-            new List<IFilterMetadata>(),
-            new Dictionary<string, object>(),
-            new object()
-        );
-
-        attribute.OnActionExecuting(actionExecutingContext);
-
-        Assert.Null(actionExecutingContext.Result);
-    }
-
-    [Fact]
-    public void ValidateRequestFilter_Validates_Request_Forbid()
-    {
-        var attribute = new ValidateRequestAttribute.InternalValidateRequestAttribute(
-            authToken: "bad",
-            null,
-            false
-        );
-
-        var form = new FormCollection(new Dictionary<string, StringValues>
-        {
-            { "key1", "value1" },
-            { "key2", "value2" }
-        });
-        var fakeContext = new ContextMocks(true, form).HttpContext.Object;
-
-        var actionExecutingContext = new ActionExecutingContext(
-            new ActionContext(fakeContext, new RouteData(), new ActionDescriptor()),
-            new List<IFilterMetadata>(),
-            new Dictionary<string, object>(),
-            new object()
-        );
-
-        attribute.OnActionExecuting(actionExecutingContext);
-
-        var statusCodeResult = (StatusCodeResult)actionExecutingContext.Result!;
-        Assert.NotNull(statusCodeResult);
-        Assert.Equal((int)HttpStatusCode.Forbidden, statusCodeResult.StatusCode);
     }
 }
