+++ src/Twilio.AspNet.Core/ValidateRequestAttribute.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Net;
+﻿using System.Net;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.DependencyInjection;
@@ -10,68 +9,31 @@ namespace Twilio.AspNet.Core
     /// <summary>
     /// Represents an attribute that is used to prevent forgery of a request.
     /// </summary>
-    public class ValidateRequestAttribute : Attribute, IFilterFactory
+    public class ValidateRequestAttribute : ActionFilterAttribute
     {
-        public bool IsReusable => false;
-
-        /// <summary>
-        /// Initializes a new instance of the ValidateRequestAttribute class.
-        /// </summary>
-        public ValidateRequestAttribute()
+        public override void OnActionExecuting(ActionExecutingContext filterContext)
         {
-        }
+            var context = filterContext.HttpContext;
+            var options = context.RequestServices
+                .GetRequiredService<IOptionsSnapshot<TwilioRequestValidationOptions>>().Value;
+            
+            var authToken = options.AuthToken;
+            var baseUrlOverride = options.BaseUrlOverride;
+            var allowLocal = options.AllowLocal ?? true;
 
-        public IFilterMetadata CreateInstance(IServiceProvider serviceProvider)
-        {
-            var options = serviceProvider.GetService<IOptionsSnapshot<TwilioRequestValidationOptions>>()?.Value;
-            if (options == null) throw new Exception("RequestValidationOptions is not configured.");
-
-            return new InternalValidateRequestAttribute(
-                authToken: options.AuthToken,
-                baseUrlOverride: options.BaseUrlOverride?.TrimEnd('/'),
-                allowLocal: options.AllowLocal ?? true
-            );
-        }
-
-        internal class InternalValidateRequestAttribute : ActionFilterAttribute
-        {
-            internal string AuthToken { get; set; }
-            internal string BaseUrlOverride { get; set; }
-            internal bool AllowLocal { get; set; }
-
-            /// <summary>
-            /// Initializes a new instance of the ValidateRequestAttribute class.
-            /// </summary>
-            /// <param name="authToken">AuthToken for the account used to sign the request</param>
-            /// <param name="baseUrlOverride">
-            /// The Base URL (protocol + hostname) to use for validation,
-            /// if different from Request.Url (sometimes needed if web site is behind a proxy or load-balancer)
-            /// </param>
-            /// <param name="allowLocal">Skip validation for local requests</param>
-            public InternalValidateRequestAttribute(string authToken, string baseUrlOverride, bool allowLocal)
+            var request = context.Request;
+            string urlOverride = null;
+            if (baseUrlOverride != null)
             {
-                AuthToken = authToken;
-                BaseUrlOverride = baseUrlOverride;
-                AllowLocal = allowLocal;
+                urlOverride = $"{baseUrlOverride}{request.Path}{request.QueryString}";
             }
 
-            public override void OnActionExecuting(ActionExecutingContext filterContext)
+            if (!RequestValidationHelper.IsValidRequest(context, authToken, urlOverride, allowLocal))
             {
-                var httpContext = filterContext.HttpContext;
-                var request = httpContext.Request;
-                string urlOverride = null;
-                if (BaseUrlOverride != null)
-                {
-                    urlOverride = $"{BaseUrlOverride}{request.Path}{request.QueryString}";
-                }
-                
-                if (!RequestValidationHelper.IsValidRequest(httpContext, AuthToken, urlOverride, AllowLocal))
-                {
-                    filterContext.Result = new StatusCodeResult((int)HttpStatusCode.Forbidden);
-                }
-
-                base.OnActionExecuting(filterContext);
+                filterContext.Result = new StatusCodeResult((int)HttpStatusCode.Forbidden);
             }
+
+            base.OnActionExecuting(filterContext);
         }
     }
 }
